# docs/swagger.yaml
basePath: /
definitions:
  ant_internal_platform_render.Response:
    properties:
      data: {}
      error:
        type: string
      status:
        type: integer
    type: object
...
  internal_product.AssignedAttribute:
    properties:
      attribute_id:
        type: integer
      is_mandatory:
        type: boolean
      name:
        type: string
      options:
        items:
          $ref: '#/definitions/internal_product.AttributeOption'
        type: array
    type: object
  internal_product.AttributeAssignmentRequest:
    properties:
      attribute_id:
        type: integer
      is_mandatory:
        type: boolean
      options:
        items:
          $ref: '#/definitions/internal_product.AttributeOptionRequest'
        type: array
    required:
    - attribute_id
    type: object
  internal_product.AttributeOption:
    properties:
      id:
        type: integer
      price_delta:
        type: number
      value:
        type: string
    type: object
  internal_product.AttributeOptionRequest:
    properties:
      option_id:
        type: integer
      price_delta:
        type: number
    required:
    - option_id
    type: object
  internal_product.CreateProductRequest:
    properties:
      attributes:
        items:
          $ref: '#/definitions/internal_product.AttributeAssignmentRequest'
        type: array
      name:
        maxLength: 200
        type: string
      price:
        minimum: 0
        type: number
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - name
    type: object
  internal_product.Product:
    properties:
      app_id:
        type: integer
      attributes:
        items:
          $ref: '#/definitions/internal_product.AssignedAttribute'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      status:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  internal_product.UpdateProductRequest:
    properties:
      attributes:
        items:
          $ref: '#/definitions/internal_product.AttributeAssignmentRequest'
        type: array
      name:
        maxLength: 200
        type: string
      price:
        minimum: 0
        type: number
      status:
        enum:
        - 0
        - 1
        type: integer
    required:
    - name
    - status
    type: object
host: localhost:8082
info:
  contact: {}
  description: This is the ant microservice.
  title: Ant API
  version: "1.0"
paths:
...
  /products:
    get:
      description: Get a list of all products for the authenticated app
      parameters:
      - description: Max items to return (default 50, max 500)
        in: query
        name: limit
        type: integer
      - description: Items to skip (default 0)
        in: query
        name: offset
        type: integer
      - description: Filter by status (0 or 1)
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_product.Product'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: List all products
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Create a new product with attribute assignments. Only active attributes
        can be assigned.
      parameters:
      - description: Product object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_product.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_product.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Create a new product
      tags:
      - products
  /products/{id}:
    delete:
      description: Delete a product and its attribute assignments. Blocked if the
        product is used in any order.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Delete product by ID
      tags:
      - products
    get:
      description: Get a single product with its assigned attributes and their options
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_product.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Get product by ID
      tags:
      - products
    put:
      consumes:
      - application/json
      description: Update an existing product. The attributes array replaces all assignments
        (full sync).
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: integer
      - description: Product object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_product.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ant_internal_platform_render.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_product.Product'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ant_internal_platform_render.Response'
      security:
      - Bearer: []
      summary: Update product by ID
      tags:
      - products
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...